Rename `chat_video_view` to `chat_view`
